add NetworkFacade and Home
