+++ src/network/NetworkFacade.ts
@@ -1,0 +1,16 @@
+import axios from 'axios'
+
+class NetworkFacade {
+  static get(url: string) {
+    return axios({
+      method: 'get',
+      url: url,
+      responseType: 'json'
+    })
+    .then(response => {
+      return response
+    })
+  }
+}
+
+export default NetworkFacade
